+++ viikko3/PDO-harjoitus/operations/selectData.php
@@ -5,6 +5,11 @@ global $SITE_URL;
 require_once __DIR__ . '/../config/config.php';
 require_once __DIR__ . '/../db/dbConnect.php';
 
+if (!isset($_SESSION['user'])) {
+    header('Location: ' . $SITE_URL . '/user.php');
+    exit;
+}
+
 $sql = "SELECT MediaItems.*, Users.username FROM MediaItems JOIN Users ON Users.user_id = MediaItems.user_id;";
 
 try {
@@ -19,8 +24,10 @@ try {
             <td><?php echo $row['username']; ?></td>
             <td><img src="./uploads/<?php echo $row['filename']; ?>" alt="<?php echo $row['title']; ?>"></td>
             <td>
-                <button class="modify-button" data-media_id="<?php echo $row['media_id']; ?>">Modify</button>
-                <a href="<?php echo $SITE_URL; ?>/operations/deleteData.php?media_id=<?php echo $row['media_id']; ?>">Delete</a>
+                <?php if ($_SESSION['user']['user_id'] == $row['user_id'] || $_SESSION['user']['user_level_id'] == 1): ?>
+                    <button class="modify-button" data-media_id="<?php echo $row['media_id']; ?>">Modify</button>
+                    <a href="<?php echo $SITE_URL; ?>/operations/deleteData.php?media_id=<?php echo $row['media_id']; ?>">Delete</a>
+                <?php endif; ?>
             </td>
         </tr>
     <?php
